requirements-tox.txt
APScheduler==3.3.1
bumpversion==0.5.3
bunch==1.0.1
configobj==5.0.6
coverage==4.5.1
django-compressor==1.5
django-cors-headers==1.1.0
django-debug-toolbar==1.6
django-extensions==1.6.7
django-gravatar2==1.1.4
django-redis==4.4.3
django-redis-metrics==1.4.0
djangorestframework==3.5.3
django-rest-swagger==2.1.1
django-solo==1.1.0
docopt==0.6.2
ecdsa==0.10
#0
-e git+https://github.com/docker/docker-py.git@aa19d7b6609c6676e8258f6b900dea2eda1dbe95#egg=docker_py-master
-e git+https://github.com/rpalacios/django-sequence-field.git#egg=django_sequence_field-master
#
gevent==1.2.2
greenlet==0.4.5
gunicorn==19.6.0
jsonfield==0.9.22
Markdown==2.6.6
mock==1.0.1
ndg-httpsclient==0.3.2
newrelic==2.94.0.79
opbeat==3.1.1
pika==0.11.0
psutil==3.3.0
psycopg2==2.7.4
pyasn1==0.1.8
pyflakes==0.8.1
Pygments==2.1.3
PyJWT==1.6.1
pyOpenSSL==16.2.0
python-jose==1.1.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
pytz==2016.7
rcssmin==1.0.6
redis==2.10.5
requests==2.13.0
sh==1.9
shortuuid==0.4
SQLAlchemy==1.1.0b1
tabulate==0.7.5
#docker-compose==1.9.0
djangorestframework-jsonp==1.0.2
pysqlite==2.8.3
PyYAML==3.12
kubernetes==3.0.0
j2cli==0.3.1.post0
tox==2.9.1
selenium==3.11.0
gitpython==2.1.9
ipaddress==1.0.22
